Adapt tests to final behavior.
HTML variables are now display with a span with the id correctly placed.

Added comment support with ~~. It is multiline but never forget to open and close it.

Try to fix as many bugs I found.

Do not know with but valdring is nor working with my linux distro with now so I can not check the lasts commits leaks. I tried to free everything thou.

Everything working.